drm/i915: Avoid integer overflow in data rate calculation
We need to upcast our plane data rate calculations to 64-bit since we're
working with numbers in 16.16 form.  Before we shift the bits off, we
could potentially have 48-bit values.

HSD: 1604167939

+++ drivers/gpu/drm/i915/intel_pm.c
@@ -3028,7 +3028,7 @@ skl_plane_relative_data_rate(const struct intel_crtc_state *cstate,
 
 	down_scale_amount = skl_plane_downscale_amount(intel_pstate);
 
-	return data_rate * down_scale_amount >> 16;
+	return (uint64_t)data_rate * down_scale_amount >> 16;
 }
 
 /*
